requirements.txt
asgiref==3.8.1
boto3==1.21.38
botocore==1.24.38
crispy-bootstrap4==2024.1
Django==5.0.6
django-crispy-forms==2.1
django-environ==0.8.1
django-ratelimit==3.0.1
django-storages==1.12.3
djangorestframework==3.15.1
gunicorn==20.1.0
jmespath==1.0.0
Pillow==9.0.1
psycopg2==2.9.9
psycopg2-binary==2.9.3
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
